+++ app/src/main/java/com/rnk0085/selfcare/ui/screen/reflection/section/moodSelector/MoodSelectorSection.kt
@@ -1,4 +1,4 @@
-package com.rnk0085.selfcare.ui.screen.reflection.component
+package com.rnk0085.selfcare.ui.screen.reflection.section.moodSelector
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -8,16 +8,12 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.rnk0085.selfcare.R
@@ -26,7 +22,7 @@ import com.rnk0085.selfcare.ui.theme.Spacing
 import com.rnk0085.selfcare.ui.theme.sectionContainerColor
 
 @Composable
-internal fun MoodSelector(
+internal fun MoodSelectorSection(
     selectedMood: MoodType?,
     onMoodSelected: (MoodType) -> Unit,
     modifier: Modifier = Modifier,
@@ -61,38 +57,11 @@ internal fun MoodSelector(
     }
 }
 
-enum class MoodType(val resourceId: Int) {
-    VERY_GOOD(R.drawable.baseline_sentiment_very_satisfied),
-    GOOD(R.drawable.baseline_sentiment_satisfied_alt),
-    NORMAL(R.drawable.baseline_sentiment_neutral),
-    BAD(R.drawable.baseline_sentiment_dissatisfied),
-    VERY_BAD(R.drawable.baseline_sentiment_very_dissatisfied),
-}
-
-@Composable
-private fun MoodIconButton(
-    moodType: MoodType,
-    isSelected: Boolean,
-    onClick: () -> Unit,
-) {
-    val iconColor = if (isSelected) Color.Black else Color.LightGray
-
-    // TODO: アイコンサイズを大きくする
-    // TODO: アイコンごとに色を設定する
-    IconButton(onClick) {
-        Icon(
-            painter = painterResource(moodType.resourceId),
-            contentDescription = null,
-            tint = iconColor,
-        )
-    }
-}
-
 @Preview
 @Composable
 private fun MoodRadioGroupPreview() {
     SelfcareTheme {
-        MoodSelector(
+        MoodSelectorSection(
             selectedMood = MoodType.GOOD,
             onMoodSelected = {},
         )
